made opactity.premodit.ioopa and moved premodit-specific io in opacity.io.ioopa to this file
